test: e2e-identity: move handle_callback to server_api.rs and drop oidc.rs
