+++ e2e-identity/tests/utils/wire_server/server_api.rs
@@ -4,7 +4,7 @@ use hyper::body::{Bytes, Incoming};
 use hyper::{Method, Request, Response, StatusCode};
 use rusty_jwt_tools::prelude::*;
 
-use crate::utils::{ctx::ctx_get, rand_base64_str, wire_server::oidc::handle_callback};
+use crate::utils::{ctx::ctx_get, rand_base64_str};
 
 // simulates wire-server database
 static mut PREVIOUS_NONCE: &str = "";
@@ -57,6 +57,23 @@ pub async fn wire_api(req: Request<Incoming>) -> http::Result<Response<Full<Byte
         (&Method::GET, ["callback"]) => handle_callback(req).await?,
         _ => not_found()?,
     })
+}
+
+async fn handle_callback(req: Request<Incoming>) -> http::Result<Response<Full<Bytes>>> {
+    let req_uri = req.uri().clone();
+    let req_uri: url::Url = format!("http://localhost{req_uri}").parse().unwrap();
+    let authorization_code = req_uri
+        .query_pairs()
+        .find_map(|(k, v)| match k.as_ref() {
+            "code" => Some(v.to_string()),
+            _ => None,
+        })
+        .unwrap();
+    let resp = Response::builder()
+        .status(StatusCode::OK)
+        .body(authorization_code.into())
+        .unwrap();
+    Ok(resp)
 }
 
 fn generate_access_token(dpop: &str, client_id: ClientId, nonce: BackendNonce) -> serde_json::Value {
